graphql-recommendation-api/requirements.txt
# GraphQL Framework
strawberry-graphql[debug-server]==0.219.0
strawberry-graphql[asgi]==0.219.0
uvicorn[standard]==0.24.0
fastapi==0.104.1

# Async support
asyncio==3.4.3
aioredis==2.0.1
aiofiles==23.2.1
httpx==0.25.2

# Database and ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Recommendation Engine
numpy==1.24.4
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2
implicit==0.7.2
surprise==1.1.3

# Vector Database
pinecone-client==2.2.4
weaviate-client==3.25.3
qdrant-client==1.6.9

# Caching
redis==5.0.1
redis-py-cluster==2.1.3

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Rate Limiting
slowapi==0.1.9
limits==3.6.0

# WebSocket support
websockets==12.0

# Monitoring & Metrics
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0

# DataLoader
aiodataloader==0.4.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
click==8.1.7
typer==0.9.0
rich==13.7.0